Fix low precision for DateTimeDataPoint in diagram
Fixed a bug where multiple data points of a series would get the same X value in the diagram.

+++ RealTimeGraphX/DataPoints/DateTimeDataPoint.cs
@@ -118,7 +118,7 @@ namespace RealTimeGraphX.DataPoints
                 return dMin.Ticks;
             }
 
-            var result = ((Value.Ticks - dMin.Ticks) * 100) / (dMax.Ticks - dMin.Ticks);
+            var result = ((Value.Ticks - dMin.Ticks) * 100) / (double)(dMax.Ticks - dMin.Ticks);
 
             return double.IsNaN(result) ? dMin.Ticks : result;
         }
